Finished separating all type of sensors apart

+++ src/demos/robot/demo_ROBOT_CRMVisualMatOpt_.cpp
@@ -366,7 +366,7 @@ int main(int argc, char* argv[]) {
 
     auto regolith_material = chrono_types::make_shared<ChVisualMaterial>();
         if (use_hapke) {
-        regolith_material->SetBSDF((unsigned int)BSDFType::HAPKE);
+        regolith_material->SetBSDF(BSDFType::HAPKE);
         regolith_material->SetHapkeParameters(0.32357f, 0.23955f, 0.30452f, 1.80238f, 0.07145f, 0.3f, 23.4f * (CH_PI / 180));
     }
     else {
