+++ src/demos/robot/viper/demo_ROBOT_Viper_SCM_WheelSinkage.cpp
@@ -287,9 +287,9 @@ int main(int argc, char* argv[]) {
 	auto regolith_mat = chrono_types::make_shared<ChVisualMaterial>();
 	auto regolith_tex_mat = chrono_types::make_shared<ChVisualMaterial>();
 	if (enable_hapke) {
-        regolith_mat->SetBSDF((unsigned int)BSDFType::HAPKE);
+        regolith_mat->SetBSDF(BSDFType::HAPKE);
         regolith_mat->SetHapkeParameters(0.32357f, 0.23955f, 0.30452f, 1.80238f, 0.07145f, 0.3f, 23.4f * (CH_PI / 180));
-		regolith_tex_mat->SetBSDF((unsigned int)BSDFType::HAPKE);
+		regolith_tex_mat->SetBSDF(BSDFType::HAPKE);
         regolith_tex_mat->SetHapkeParameters(0.32357f, 0.23955f, 0.30452f, 1.80238f, 0.07145f, 0.3f, 23.4f * (CH_PI / 180));
     }
     else {
